feat: flightList as a component, origin/destination style fix
