+++ src/basket.json
@@ -1,1 +1,12 @@
-[]
+[
+  {
+    "date": "2024-02-26T11:16:26.041Z",
+    "uuid": "cd679907-ddae-4f24-b62e-f6fc3118765e",
+    "data": [
+      {
+        "firstName": "asd",
+        "lastName": "dfg"
+      }
+    ]
+  }
+]
